correct translations and rename dlatch to dflipflop
