+++ assets/levels/level_15.json
@@ -88,16 +88,12 @@
         }
     ],
     "hints": [
-        "Use 4 of the inputs as the data inputs (A,B,C,D) for the 4-input MUX and combine the remaining 2 inputs with the Collector to form a 2-bit selector.",
-        "A typical mapping: Inputs 0-3 = data inputs A..D, Inputs 4-5 = selector bits S0 (LSB), S1 (MSB).",
-        "The collector will pack the two selector bits into a 2-bit control signal for the 4-input MUX.",
-        "The expression for a 4-to-1 multiplexer is: $(A \\cdot \\overline{S1} \\cdot \\overline{S0}) + (B \\cdot \\overline{S1} \\cdot S0) + (C \\cdot S1 \\cdot \\overline{S0}) + (D \\cdot S1 \\cdot S0)$."
+        "Think about how you built the 2-input MUX. Use the same principle but with two control signals this time",
+        "This is a bonus level, so you have to think for yourself now!"
     ],
     "hints_de": [
-        "Verwende 4 der Eingänge als Daten (A,B,C,D) für den 4-Eingang-MUX und kombiniere die verbleibenden 2 Eingänge mit dem Kollektor, um einen 2-Bit-Selektor zu bilden.",
-        "Eine typische Zuordnung: Eingänge 0-3 = Daten A..D, Eingänge 4-5 = Selektor-Bits S0 (LSB), S1 (MSB).",
-        "Der Kollektor packt die zwei Selektor-Bits in ein 2-Bit-Steuersignal für den 4-Eingang-MUX.",
-        "Der Ausdruck für einen 4-zu-1-Multiplexer lautet: $(A \\cdot \\overline{S1} \\cdot \\overline{S0}) + (B \\cdot \\overline{S1} \\cdot S0) +$\n$(C \\cdot S1 \\cdot \\overline{S0}) + (D \\cdot S1 \\cdot S0)$."
+        "Denke darüber nach, wie du den 2-Eingangs-MUX gebaut hast. Verwende dasselbe Prinzip, aber diesmal mit zwei Steuersignalen",
+        "Dies ist ein Bonuslevel, also musst du jetzt selbst nachdenken!"
     ],
     "tests": [
         {
