+++ assets/levels/level_29.json
@@ -558,7 +558,7 @@
         }
     ],
     "name_de": "1-Bit-ALU",
-    "description_de": "Wir sind jetzt im Prozess, einen Prozessor zusammenzustellen. Die tatsächliche Berechnung erfolgt in der Arithmetic-Logical-Unit (ALU). In dieser Stufe bauen wir eine einfache 1-Bit-ALU, die A UND B, A ODER B oder A + B (Summenbit) berechnen kann. Ein 2-Bit-Steuersignal `op` wählt die Operation. Später wollen wir die ALU auf mehr Bits erweitern, also muss unser Addierer einen Carry-In und einen Carry-Out haben. Verwende für jetzt einen separaten \"Dummy\"-Eingang als Carry-In mit 0 verbunden und verbinde den Carry-Out mit dem gekennzeichneten Ausgang.",
+    "description_de": "Wir befassen uns jetzt damit, einen Prozessor zusammenzustellen. Die tatsächliche Berechnung erfolgt in der Arithmetic-Logical-Unit (ALU). In dieser Stufe bauen wir eine einfache 1-Bit-ALU, die A UND B, A ODER B oder A + B (Summenbit) berechnen kann. Ein 2-Bit-Steuersignal `op` wählt die Operation. Später wollen wir die ALU auf mehr Bits erweitern, also muss unser Addierer einen Carry-In und einen Carry-Out haben. Verwende für jetzt einen separaten \"Dummy\"-Eingang als Carry-In mit 0 verbunden und verbinde den Carry-Out mit dem gekennzeichneten Ausgang.",
     "difficulty_de": "Mittel",
     "objectives_de": [
         "Baue eine 1-Bit-ALU mit Eingängen A und B und einem 2-Bit-Operationscode `op` (op1,op0).",
