+++ lib/core/components/component_registry.dart
@@ -21,7 +21,7 @@ import 'package:circuitquest/core/components/gates/or_gate.dart';
 import 'package:circuitquest/core/components/gates/xor_gate.dart';
 import 'package:circuitquest/core/components/input_source.dart';
 import 'package:circuitquest/core/components/output_probe.dart';
-import 'package:circuitquest/core/components/sequential/d_latch.dart';
+import 'package:circuitquest/core/components/sequential/d_flip_flop.dart';
 import 'package:circuitquest/core/components/sequential/register.dart';
 
 class ComponentType {
@@ -103,7 +103,7 @@ final List<ComponentType> availableComponents = [
   ),
   ComponentType(
     name: 'DFlipflop',
-    displayName: 'D-Latch',
+    displayName: 'D-Flipflop',
     iconPath: 'assets/gates/DFlipflop.svg',
     createComponent: () => DFlipflop(),
   ),
